Revert "Merge branch 'merge_upstream'"
This reverts commit 7a5c1bcab9d4091ce8c671f94fb89600232fc7fd, reversing
changes made to 2cd6bff7721cdcb71f85c09f34ae96e0345ed392.

+++ config/config.align.c
@@ -7,14 +7,13 @@
   notice and this notice are preserved.
 
 */
-#include <stdlib.h>
 
 int main ()
 {
   char  *buf = malloc(30);
   void  *v;
-  short *sp;
-  short *sq;
+  int   *sp;
+  int   *sq;
   int   *ip;
   int   *iq;
   int   i;
